kernel32: Read entry before LdrInitializeThunk
Overwatch overwrites the PE header contents in a TLS callback. This results in
a crash on wine, because the entry point will be incorrect in start_process.

Fix the issue by reading the entry point before LdrInitializeThunk is called,
and using this stored value in start_process instead of the PE header.

Signed-off-by: Andrew Wesie <awesie@gmail.com>