ntdll: Overwatch compatibility for ntdll
Overwatch requires that it can load the ntdll.dll file into memory as RWX,
and then run Nt* functions directly from it. This works on normal Windows
because those functions are system calls.

It additionally verifies that some of the functions in the in-memory
ntdll.dll (loaded by wine) and the on-disk ntdll.dll start with the same
sequence of bytes.

The current solution is a hack. We force ntdll.dll to be compiled without
optimizations so there is a known sequence of instructions at the beginning
of most functions. We construct a fake ntdll.dll with trampolines that start
with the same sequence of instructions and then jump into the real ntdll.dll
functions.

The fixup script is not currently included in the build process, so the user
must run it manually after build. The user should copy the template dll into
their system32 directory, and then run the fixup script with the path to
ntdll.dll.so and system32\ntdll.dll.

Signed-off-by: Andrew Wesie <awesie@gmail.com>